Update color parsing
