+++ pyvista/plotting/__init__.py
@@ -1,5 +1,5 @@
 from .colors import (color_char_to_word, get_cmap_safe, hex_to_rgb, hexcolors,
-                     string_to_rgb)
+                     string_to_rgb, parse_color)
 from .export_vtkjs import export_plotter_vtkjs, get_vtkjs_url
 from .export_x3d import export_x3d, X3D_JAVASCRIPT
 from .export_k3d import export_k3d
@@ -10,6 +10,6 @@ from .plotting import BasePlotter, Plotter, close_all
 from .qt_plotting import BackgroundPlotter, QtInteractor
 from .renderer import Renderer
 from .theme import (DEFAULT_THEME, FONT_KEYS, MAX_N_COLOR_BARS, PV_BACKGROUND,
-                    parse_color, parse_font_family, rcParams, set_plot_theme)
+                    parse_font_family, rcParams, set_plot_theme)
 from .tools import (create_axes_orientation_box, opacity_transfer_function,
                     system_supports_plotting)
